2.6.1- added default values, choice values
"default"=>'' This takes a comma delimited list and places the values in the fields of an insert form as the default value of the                   field.  
      default="today,,,,'
"choice_vals"=>''  This takes in the values for the choice field type (see below)
Field Type of choice added.  define it as part of the element style syntax and then add the choice vals with the format of
              choice_vals=field^value1^value2^value3!field^value1^value2^value3